Enhance user registration and UI styling
Added a `ConfirmPassword` property to the `UserRegistrationDTO` class to ensure password confirmation matches the password. Updated `Registeration.cshtml` to include fields for `ConfirmPassword` and fixed a typo in the `@model` directive. Removed an unnecessary closing brace and `@section Scripts` from `DeleteUserForm.cshtml`.

Improved the layout and styling of `CharacterDetailsView.cshtml` and `WeaponDetailsView.cshtml` by using a hero section with an overlay instead of a card. Enhanced the visual styling of `CharacterListView.cshtml` by adding a container with a background and adjusting text and hr elements to be white.

Made several CSS changes in `site.css` to improve the visual styling, including adjustments to the body margin, button focus styles, hero section background, and card hover effects. Added new styles for form controls, buttons, and alerts to enhance the user interface.

Added a new migration `20250323141227_AddTypeToCharacter.cs` to include a `Type` column in the `Characters` table in the database.